Week-2/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
cryptography==44.0.0
fastapi==0.115.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
locust==2.32.8
MarkupSafe==3.0.2
msgpack==1.1.0
psutil==6.1.1
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
pyzmq==26.2.1
redis==5.2.1
requests==2.32.3
setuptools==75.8.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
